XrdCephPosix: Fix 'Offline' file status
XRootD assumes a file is 'Offline' if the st_dev and st_ino
members returned via a stat are set to zero (defaults).
Workflows may use the 'Offline' flag, so it is important
that it is correct. Since Ceph object store is disk-based,
it is sufficient to default st_dev and st_ino to non-zero
values, such that XRootD determines a file to be 'Online';
the hashes of pool and file names are used to calculate more
meaningful values.

Note:
 - The values used for st_dev and st_ino are not necessarily
   those returned by the hash function, due to the type
   conversion.

See:
https://github.com/xrootd/xrootd/blob/master/src/XrdSfs/XrdSfsFlags.hh#L96